# contracts/api/llm-gateway/v1/openapi.yaml
openapi: 3.0.3
info:
  title: LoreWeave LLM Gateway API
  version: 1.0.0
  description: |
    Unified LLM contract for the LoreWeave platform. Two flavors:

    1. **`POST /v1/llm/stream`** — interactive streaming chat (P1).
       Server-Sent Events end-to-end. **No wall-clock timeout.**
       Token-level deltas reach the caller as the upstream provider
       emits them.

    2. **`POST /v1/llm/jobs`** — fire-and-forget async LLM job (P2).
       Returns 202 with `job_id` immediately. Job runs to completion
       on the gateway side (with chunking, parallel chunk dispatch,
       and aggregation). Completion is signaled out-of-band via the
       caller-supplied callback (RabbitMQ topic or webhook URL); the
       caller never blocks on HTTP for the result. `GET /v1/llm/jobs/{id}`
       is for status polling; `DELETE /v1/llm/jobs/{id}` cancels.

    Every other LLM call site in the monorepo MUST use one of these
    two endpoints (or their `/internal/llm/*` service-to-service
    counterparts). Direct provider-SDK calls (`litellm`, `openai`,
    `anthropic`) are forbidden by the SDK lint layer.

    See [LLM_PIPELINE_UNIFIED_REFACTOR_PLAN.md](../../../docs/03_planning/LLM_PIPELINE_UNIFIED_REFACTOR_PLAN.md)
    for principles, audit findings, and migration phases.
servers:
  - url: http://localhost:3000
    description: API Gateway (BFF) for browser callers
  - url: http://provider-registry-service:8085
...
paths:
...
  /v1/llm/jobs:
    post:
      tags: [jobs]
      summary: Submit an async LLM job
      description: |
        Returns 202 with `job_id` immediately. The job runs on the gateway
        with optional chunking + parallel dispatch + aggregation.

        Completion is signaled via the caller-supplied callback:
        - **RabbitMQ topic** (`callback.kind = "rabbitmq"`): an event is
          published to `routing_key` on the platform's `loreweave.events`
          topic exchange. Preferred for in-cluster consumers.
        - **Webhook URL** (`callback.kind = "webhook"`): an HTTP POST to
          `url` is made with the completed `Job` body. Webhook payloads
          are signed with HMAC-SHA256 (header `X-LLM-Signature`).

        The caller may also poll `GET /v1/llm/jobs/{id}` if no callback
        is registered (callbacks are optional).

        **Spend guardrail (Phase 6a).** Before the job is accepted the
        gateway places a USD reservation against the caller's budget:
        - `404 LLM_MODEL_NOT_FOUND` — `model_ref` resolves to no model.
        - `402 LLM_QUOTA_EXCEEDED` — the model exists but has no pricing
          configured; or the worst-case cost estimate exceeds the
          caller's remaining daily/monthly budget; or, for a
          `platform_model` job, it exceeds the caller's LoreWeave free
          tier + prepaid credits (Phase 6a-β Subsystem B).
        For a `chat`/`completion` job that would only marginally exceed
        budget, the gateway instead caps `max_tokens` to what the budget
        affords and records the cap under `job_meta.max_tokens_capped`
        (`{requested, applied, reason}`) — the result is never silently
        truncated.
      operationId: submitLlmJob
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitJobRequest'
          multipart/form-data:
            # Phase 5b — bytes-mode STT submission. Audio carried inline
            # via multipart file part; metadata in form fields. Only
            # operation=stt is accepted via multipart. The handler
            # uses mime.ParseMediaType to dispatch (case-insensitive
            # per RFC 2045).
            schema:
              $ref: '#/components/schemas/SubmitSttBytesRequest'
      responses:
        '202':
          description: Job accepted. Result available later via callback or polling.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitJobResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '402':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          # Phase 5b — multipart audio exceeds 25MB cap.
          $ref: '#/components/responses/ErrorResponse'
        '415':
          # Phase 5b — Content-Type neither application/json nor multipart/form-data.
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
  /internal/llm/jobs:
    post:
      tags: [jobs-internal]
      summary: Service-to-service async LLM job submission
      operationId: submitInternalLlmJob
      security:
        - internalServiceToken: []
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitJobRequest'
          multipart/form-data:
            # Phase 5b — bytes-mode STT. Same shape as /v1/llm/jobs;
            # both share doSubmitJob handler.
            schema:
              $ref: '#/components/schemas/SubmitSttBytesRequest'
      responses:
        '202':
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitJobResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '402':
          # Phase 6a — spend guardrail: unpriced model or over budget.
          $ref: '#/components/responses/ErrorResponse'
        '404':
          # Phase 6a — model_ref resolves to no model.
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '415':
          $ref: '#/components/responses/ErrorResponse'
  /v1/llm/jobs/{job_id}:
    get:
      tags: [jobs]
      summary: Get LLM job status / result
      description: |
        Returns the current state of a job. Until terminal, `result` and
        `error` are null. Once terminal (status ∈ {completed, failed,
        cancelled}), the appropriate field is populated.

        Result payload retention defaults to 7 days post-completion;
        polling after that returns 404.
      operationId: getLlmJob
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'

    delete:
      tags: [jobs]
      summary: Cancel an in-flight LLM job
      description: |
        Best-effort cancel. The gateway sends cancellation to the upstream
        provider where supported. In-flight chunks may still complete and
        be billed (provider-dependent). Returns 204 if the cancel was
        accepted; the job's final status is reflected in subsequent GETs.
      operationId: cancelLlmJob
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '204':
          description: Cancel accepted
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          description: Job already terminal (completed/failed/cancelled) — no-op
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /internal/llm/jobs/{job_id}:
    get:
      tags: [jobs-internal]
      summary: Service-to-service job status fetch
      operationId: getInternalLlmJob
      security:
        - internalServiceToken: []
      parameters:
        - $ref: '#/components/parameters/JobId'
        - name: user_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalServiceToken:
      type: apiKey
      in: header
      name: X-Internal-Token
  parameters:
    JobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ErrorResponse:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
...
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            Stable error code. Common values:
            `LLM_INVALID_REQUEST`, `LLM_AUTH_FAILED`, `LLM_QUOTA_EXCEEDED`,
            `LLM_MODEL_NOT_FOUND`, `LLM_RATE_LIMITED`, `LLM_UPSTREAM_ERROR`,
            `LLM_DECODE_ERROR`, `LLM_BODY_TOO_LARGE`, `LLM_JOB_NOT_FOUND`,
            `LLM_JOB_TERMINAL`.
        message:
          type: string
        retry_after_s:
          type: number
          description: Present on `LLM_RATE_LIMITED` (and `429` responses).
          nullable: true

    # NOTE: `ProviderKind` is defined in the model-registry contract
    # and resolved server-side from `model_ref`. Callers of this gateway
    # never specify it directly, so it's not a schema in this contract.
    ModelSource:
      type: string
      enum: [user_model, platform_model]

    # ── Streaming (P1) ────────────────────────────────────────────────
...
    JobOperation:
      type: string
      enum:
        - chat
        - completion
        - embedding
        - stt
        - tts
        - image_gen
        - video_gen
        - audio_gen
        - entity_extraction
        - relation_extraction
        - event_extraction
        - fact_extraction
        - summarize_level
        - translation
        - vision
      description: |
        The semantic operation. The gateway uses this to:
        1. Pick the right adapter method per provider
        2. Pick the right per-operation chunker (e.g. `entity_extraction`
           uses paragraph-based chunking; `translation` uses token-based
           chunking with overlap)
        3. Pick the right aggregator (entity dedup-by-canonical-id;
           translation concat; chat concat; embedding stack)
    ChunkingConfig:
      type: object
      required: [strategy]
      properties:
        strategy:
          type: string
          enum: [tokens, paragraphs, sentences, none]
          default: none
          description: |
            - `tokens`: split by token count using `tiktoken cl100k_base`.
            - `paragraphs`: split on blank-line boundaries.
            - `sentences`: split with langdetect-driven sentence segmenter.
            - `none`: pass the input as-is to a single LLM call.
        size:
          type: integer
          minimum: 1
          description: |
            Target chunk size. Units depend on strategy:
            - `tokens`: token count (default 2000)
            - `paragraphs`: paragraph count (default 8)
            - `sentences`: sentence count (default 30)
        overlap:
          type: integer
          minimum: 0
          description: |
            Optional inter-chunk overlap, same units as `size`. Default 0
            for paragraphs/sentences; default 200 tokens for tokens.
    CallbackConfig:
      type: object
      required: [kind]
      properties:
        kind:
          type: string
          enum: [rabbitmq, webhook]
        routing_key:
          type: string
          description: |
            Required when `kind = rabbitmq`. Topic-exchange routing key
            published on `loreweave.events`. Convention:
            `user.<user_id>.llm.<operation>.<event>`
            Example: `user.019d5e3c-...-1344e148bf7c.llm.entity_extraction.done`
          nullable: true
        url:
          type: string
          format: uri
          description: |
            Required when `kind = webhook`. Webhook payload is the
            completed `Job` body; signed via `X-LLM-Signature: sha256=<hex>`
            using a shared secret (registered out-of-band per caller).
          nullable: true
    SubmitJobRequest:
      type: object
      required: [operation, model_source, model_ref, input]
      properties:
        operation:
          $ref: '#/components/schemas/JobOperation'
        model_source:
          $ref: '#/components/schemas/ModelSource'
        model_ref:
          type: string
          format: uuid
        input:
          type: object
          additionalProperties: true
          description: |
            Operation-specific payload.

            **Streaming-backed ops** (`chat`, `completion`, `entity_extraction`,
            `relation_extraction`, `event_extraction`, `fact_extraction`)
            ALL share the OpenAI chat-completion wire shape:
              `{ messages: [...], tools: [...], temperature, max_tokens, response_format }`
            The operation enum picks the per-op aggregator (chat = concat,
            entity_extraction = jsonListAggregator over `entities`, etc.) —
            it does NOT change the input schema. Callers that want to send
            extraction prompts route them via the `messages` field; the
            extraction op's name only signals which aggregator to apply
            to the streamed JSON output.

            **Other ops** have op-specific input shapes:
            - `embedding`: `{ texts: ["...", "..."] }`
            - `translation`: `{ text: "...", source_lang: "en", target_lang: "vi", glossary: [...] }`
            - `image_gen`: `{ prompt: "...", size: "1024x1024", n: 1 }`
            - `stt`: see `SttInput` (Phase 5a — `{ audio_url, language }`)
            - `tts`: **NOT supported via jobs** — submit via `POST /v1/llm/stream`
              with `operation: tts` instead. Submitting `tts` here returns
              `LLM_OPERATION_NOT_SUPPORTED_VIA_JOBS` with a hint pointing
              at the streaming endpoint. This is intentional: TTS is a
              UX-critical low-latency case that the SSE path serves
              naturally, and forcing it through buffered job-pattern would
              add unwanted wait time before first audio.
        chunking:
          oneOf:
            - $ref: '#/components/schemas/ChunkingConfig'
          nullable: true
        callback:
          oneOf:
            - $ref: '#/components/schemas/CallbackConfig'
          nullable: true
        trace_id:
          type: string
          nullable: true
        job_meta:
          type: object
          additionalProperties: true
          description: |
            Arbitrary caller metadata echoed back in `Job.job_meta`. Use
            for business-job linkage (e.g. `{ extraction_job_id, chapter_id }`).
          nullable: true
    SubmitJobResponse:
      type: object
      required: [job_id, status, submitted_at]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/JobStatus'
        submitted_at:
          type: string
          format: date-time
    JobStatus:
      type: string
      enum: [pending, running, completed, failed, cancelled]
    JobProgress:
      type: object
      properties:
        chunks_total:
          type: integer
          minimum: 0
          nullable: true
          description: Total chunks for the job. Null when chunking is `none`.
        chunks_done:
          type: integer
          minimum: 0
          nullable: true
        tokens_used:
          type: integer
          minimum: 0
          description: Cumulative tokens consumed across all chunks so far.
        last_progress_at:
          type: string
          format: date-time
          nullable: true
    Job:
      type: object
      required: [job_id, operation, status, submitted_at]
      properties:
        job_id:
          type: string
          format: uuid
        operation:
          $ref: '#/components/schemas/JobOperation'
        status:
          $ref: '#/components/schemas/JobStatus'
        progress:
          oneOf:
            - $ref: '#/components/schemas/JobProgress'
          nullable: true
        result:
          type: object
          additionalProperties: true
          description: |
            Operation-specific result payload, populated when
            `status = completed`. Schemas (informal):
            - `chat`: `{ messages: [{role,content}], usage: {...} }`
            - `embedding`: `{ vectors: [[...], ...], model: "..." }`
            - `entity_extraction`: `{ entities: [{name,kind,aliases,confidence}, ...] }`
            - `relation_extraction`: `{ relations: [{subject,predicate,object,confidence}, ...] }`
            - `event_extraction`: `{ events: [{summary,participants,event_date,...}, ...] }`
            - `translation`: `{ translated_text: "...", source_lang, target_lang, usage }`
            - `image_gen`: `{ images: [{url|b64,...}, ...] }`
            - `stt`: see `SttResult` (Phase 5a — `{ text, language, duration_ms }`)
            - `tts`: never populated here — TTS only streams via
              `/v1/llm/stream` and returns no Job.result.
          nullable: true
        error:
          oneOf:
            - $ref: '#/components/schemas/ErrorBody'
          nullable: true
        submitted_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        trace_id:
          type: string
          nullable: true
        job_meta:
          type: object
          additionalProperties: true
          nullable: true

    # ── Audio operation schemas (Phase 5a) ─────────────────────────────
...
    SubmitSttBytesRequest:
      type: object
      required: [operation, model_source, model_ref, audio]
      description: |
        Phase 5b. Multipart variant for STT submission with the audio
        bytes carried inline. The gateway streams the file into a
        per-process buffer (no DB persistence), dispatches the worker,
        and replies 202 with the same `SubmitJobResponse` envelope as
        the JSON variant.

        Only `operation=stt` is accepted on `multipart/form-data`. Other
        operations on multipart return 400 `LLM_INVALID_REQUEST`.

        `chunking` config is NOT accepted on multipart submits — STT
        runs as a single upstream call regardless of audio length. If a
        `chunking` form field is present, the handler returns 400
        `LLM_INVALID_REQUEST: "chunking not accepted on stt multipart submits"`.

        File field name MUST be exactly `audio`. Anything else returns
        400 `LLM_INVALID_REQUEST: "expected file field 'audio'; got [<list>]"`.
      properties:
        operation:
          type: string
          enum: [stt]
          description: Must be `stt`.
        model_source:
          $ref: '#/components/schemas/ModelSource'
        model_ref:
          type: string
          format: uuid
          description: Caller's model UUID (resolved via provider-registry).
        language:
          type: string
          default: auto
          description: |
            ISO 639-1 code or `auto` for upstream auto-detection.
            Mirrors `SttInput.language`.
        trace_id:
          type: string
          nullable: true
          description: Optional correlation ID, mirrors `SubmitJobRequest.trace_id`.
        audio:
          type: string
          format: binary
          description: |
            Audio bytes (any format Whisper accepts: webm, mp3, wav,
            ogg, m4a, flac). 25MB hard cap matches OpenAI Whisper
            upload limit. Bytes > 25MB rejected with 413
            `LLM_AUDIO_TOO_LARGE`. Filename is informational — adapter
            picks the upstream extension via Content-Type.
...